+++ src/objects/CanvasPath.ts
@@ -1,0 +1,7 @@
+import Shape from './Shape'
+
+export default class CanvasPath extends Shape {
+    constructor(x?: number, y?: number) {
+        super(x, y)
+    }
+}